requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
pathlib2==2.3.7
python-docx==0.8.11
PyPDF2==3.0.1
pdfplumber==0.10.3